add red circle and shade to matrix of heat trees
change in figure visualization to make comamonadaceae more visible in this busy plot. this was performed in inkscape. to shade the objects, first all trees and all dotted red circles were duplicated instead of copy-pasted (https://inkscape-manuals.readthedocs.io/en/latest/copy-and-duplicate.html). after that, a clip (https://inkscape-manuals.readthedocs.io/en/latest/clipping-and-masking.html) was set for each tree and it's comamonadaceae red circle. Then a 60% opaque white square was moved on top of the full tree but below the cliped comamonadace branch. adjustment of objets being on top or on different layers might was needed