ordinations of full communities and above-neutral. need proper svg verison
this figure might be supressed from the manuscript. it does how an overlap between meja 0.1mM and oral secretion, so I think Marcel is going to like it. I had an issue in saving the plot as an orbject in the R envirment, so I did not export it as an svg file with code. details for this figure can be handled later